refactor: move agents-ui integration code out of Eliza folder
