+++ src/pages/eliza.astro
@@ -1,7 +1,7 @@
 ---
 import settings from '@base/settings.json';
 // import ElizaModule from '@components/Eliza';
-import ElizaModule from '@components/Eliza/ElizaHost';
+import ElizaModule from '@components/AgentsUI';
 import BaseHtml from '@layouts/BaseHtml.astro';
 ---
